Add single file execution functionality
- A play button will now appear when viewing a RISC-V assembly file, allowing it to be run without an explicit launch.json file. If a launch file is detected, it will override the default config (same behavior as before).
- Config retrieval code has been moved to a common utilities class
- Added `ece2035.runFile` command (also executable through command palette)
