+++ ece2035/package.json
@@ -30,6 +30,11 @@
         "icon": "$(play)"
       },
       {
+          "command": "ece2035.runFile",
+          "title": "Run RISCV File",
+          "icon": "$(play)"
+      },
+      {
         "command": "ece2035.debugTestCase",
         "title": "Debug RISCV Test Case",
         "icon": "$(debug-alt)"
@@ -130,6 +135,13 @@
       ]
     },
     "menus": {
+      "editor/title/run": [
+        {
+          "command": "ece2035.runFile",
+          "when": "editorLangId == riscv && resourceExtname == .asm",
+          "group": "navigation"
+        }
+      ],
       "view/item/context": [
         {
           "command": "ece2035.deleteTestCase",
