+++ ece2035/src/defaults.ts
@@ -1,0 +1,6 @@
+export const DEFAULT_LAUNCH_CONFIG = {
+  "type": "riscv-vm",
+  "request": "launch",
+  "name": "Run File",
+  "program": "${file}",
+};
